BaseLayout 작업 완료 (기존 코드 Backup)
