+++ frontend/src/components/LoginComponent.vue
@@ -1,76 +1,54 @@
 <template>
-    <div class="div">
-      <img class="image-6" src="../assets/images/public/image-60.png" />
-      <div class="user">user</div>
-      <div class="rectangle-4"></div>
-      <div class="rectangle-5"></div>
-      <div class="rectangle-6"></div>
-      <div class="address">address</div>
-      <div class="rectangle-8"></div>
-      <img class="rectangle-9" src="../assets/images/public/rectangle-90.png" />
-      <div class="search">search</div>
-      <img class="polygon-6" src="../assets/images/public/polygon-60.png" />
-      <div class="login">login</div>
-      <img class="image-16" src="../assets/images/public/image-150.png" />
-      <div class="rectangle-1"></div>
-      <div class="rectangle-3"></div>
-      <div class="my-page">My page</div>
-      <div class="my-categories">My categories</div>
-      <div class="mappride">Mappride</div>
-      <div class="line-1"></div>
-      <div class="rectangle-2"></div>
-      <div class="map">Map</div>
-      <img class="image-12" src="../assets/images/public/image-120.png" />
-      <img class="image-13" src="../assets/images/public/image-130.png" />
-      <img class="image-14" src="../assets/images/public/image-140.png" />
-      <div class="rectangle-12"></div>
-      <div class="rectangle-13"></div>
-      <!-- <img class="web-light-sq-si-4-x" src="../assets/images/loginComponent/web-light-sq-si-4-x0.png" />
-      <img
-        class="kakao-login-medium-narrow"
-        src="../assets/images/loginComponent/kakao-login-medium-narrow0.png"
-      />
-      <img class="btn-g" src="../assets/images/loginComponent/btn-g0.png" /> -->
+  <div class="full">
+    <img class="image-6" src="../assets/images/public/image-60.png" />
+    <div class="user">user</div>
 
-      <button class="login-button">
-        <img src="../assets/images/loginComponent/web-light-sq-si-4-x0.png" alt="Web Login" />
-      </button>
-      
-      <button class="kakao-login-button">
-        <img src="../assets/images/loginComponent/kakao-login-medium-narrow0.png" alt="Kakao Login" />
-      </button>
-      
-      <button class="google-login-button">
-        <img src="../assets/images/loginComponent/btn-g0.png" alt="Google Login" />
-      </button>
+    <div class="line-1"></div>
+    <div class="rectangle-12"></div>
+    <div class="rectangle-13"></div>
+    <!-- <img class="web-light-sq-si-4-x" src="../assets/images/loginComponent/web-light-sq-si-4-x0.png" />
+    <img
+      class="kakao-login-medium-narrow"
+      src="../assets/images/loginComponent/kakao-login-medium-narrow0.png"
+    />
+    <img class="btn-g" src="../assets/images/loginComponent/btn-g0.png" /> -->
 
-      <div class="line-13"></div>
-    </div>
-  </template>
-  <script>
+    <button class="login-button">
+      <img src="../assets/images/loginComponent/web-light-sq-si-4-x0.png" alt="Web Login" />
+    </button>
+    
+    <button class="kakao-login-button">
+      <img src="../assets/images/loginComponent/kakao-login-medium-narrow0.png" alt="Kakao Login" />
+    </button>
+    
+    <button class="google-login-button">
+      <img src="../assets/images/loginComponent/btn-g0.png" alt="Google Login" />
+    </button>
+    <div class="line-13"></div>
+  </div>
+</template>
+
+<script>
   export default {
     name: "Component",
     components: {},
     props: {},
     data() {},
   };
-  </script>
-  <style scoped>
-  .div,
-  .div * {
+</script>
+  
+<style scoped>
+  .full * {
     box-sizing: border-box;
   }
-  .div {
-    background: #ffffff;
-    height: 1080px;
-    position: relative;
-    overflow: hidden;
+  .full {
+    top: 10px;
+    height: 100%;
+    width: 1600px;
   }
   .image-6 {
-    width: 1668px;
-    height: 1002px;
     position: absolute;
-    left: 252px;
+    left: 350px;
     top: 78px;
     object-fit: cover;
     aspect-ratio: 1668/1002;
@@ -86,90 +64,6 @@
     top: 77px;
     -webkit-text-stroke: 1px #d2d2d2;
   }
-  .rectangle-4 {
-    background: #ffffff;
-    width: 1669px;
-    height: 86px;
-    position: absolute;
-    left: 251px;
-    top: 0px;
-  }
-  .rectangle-5 {
-    background: #ffffff;
-    border-style: solid;
-    border-color: #d2d2d2;
-    border-width: 1px;
-    width: 460px;
-    height: 46px;
-    position: absolute;
-    left: 752px;
-    top: 26px;
-  }
-  .rectangle-6 {
-    background: #ffffff;
-    border-style: solid;
-    border-color: #d2d2d2;
-    border-width: 1px;
-    width: 165px;
-    height: 45px;
-    position: absolute;
-    left: 568px;
-    top: 27px;
-  }
-  .address {
-    color: #000000;
-    text-align: left;
-    font-family: "Stylish-Regular", sans-serif;
-    font-size: 20px;
-    font-weight: 400;
-    position: absolute;
-    left: 604px;
-    top: 36px;
-    width: 82px;
-    height: 26px;
-    -webkit-text-stroke: 1px #d2d2d2;
-  }
-  .rectangle-8 {
-    background: #ffffff;
-    border-style: solid;
-    border-color: #d2d2d2;
-    border-width: 1px;
-    width: 126px;
-    height: 45px;
-    position: absolute;
-    left: 1239px;
-    top: 27px;
-  }
-  .rectangle-9 {
-    width: 126px;
-    height: 45px;
-    position: absolute;
-    left: 1774px;
-    top: 27px;
-    overflow: visible;
-  }
-  .search {
-    color: #000000;
-    text-align: left;
-    font-family: "Stylish-Regular", sans-serif;
-    font-size: 20px;
-    font-weight: 400;
-    position: absolute;
-    left: 1267px;
-    top: 36px;
-    width: 70px;
-    height: 26px;
-    -webkit-text-stroke: 1px #d2d2d2;
-  }
-  .polygon-6 {
-    width: 23px;
-    height: 15px;
-    position: absolute;
-    left: 714.09px;
-    top: 56px;
-    transform: translate(-21.52px, -11.24px);
-    overflow: visible;
-  }
   .login {
     color: #000000;
     text-align: left;
@@ -182,87 +76,9 @@
     width: 67px;
     height: 26px;
   }
-  .image-16 {
-    width: 46px;
-    height: 46px;
-    position: absolute;
-    left: 1633px;
-    top: 18px;
-    object-fit: cover;
-    aspect-ratio: 1;
-  }
   .group-1 {
     position: absolute;
     inset: 0;
-  }
-  .rectangle-1 {
-    background: #e5e5e5;
-    border-style: solid;
-    border-color: #ffffff;
-    border-width: 0px 1px 0px 0px;
-    width: 248.38px;
-    height: 1080px;
-    position: absolute;
-    left: 1.62px;
-    top: 0px;
-    box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-  }
-  .rectangle-3 {
-    background: #ffffff;
-    border-style: solid;
-    border-color: #efefef;
-    border-width: 0px 1px 0px 0px;
-    width: 248.38px;
-    height: 1080px;
-    position: absolute;
-    left: 1.62px;
-    top: 0px;
-    box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-  }
-  .my-page {
-    color: #000000;
-    text-align: center;
-    font-family: "Stylish-Regular", sans-serif;
-    font-size: 25px;
-    font-weight: 400;
-    position: absolute;
-    left: 1.62px;
-    top: 420.82px;
-    width: 248.38px;
-    height: 104.41px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    -webkit-text-stroke: 1px #ffffff;
-  }
-  .my-categories {
-    color: #000000;
-    text-align: center;
-    font-family: "Stylish-Regular", sans-serif;
-    font-size: 25px;
-    font-weight: 400;
-    position: absolute;
-    left: 1.62px;
-    top: 495.7px;
-    width: 248.38px;
-    height: 104.41px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    -webkit-text-stroke: 1px #ffffff;
-  }
-  .mappride {
-    color: #000000;
-    text-align: left;
-    font-family: "Stylish-Regular", sans-serif;
-    font-size: 30px;
-    font-weight: 400;
-    position: absolute;
-    left: 34.72px;
-    top: 28.48px;
-    width: 137.9px;
-    height: 40.08px;
-    -webkit-text-stroke: 1px #ffffff;
   }
   .line-1 {
     margin-top: -1px;
@@ -275,67 +91,12 @@
     left: 0px;
     top: 144.49px;
   }
-  .rectangle-2 {
-    background: #f7f7f7;
-    border-style: solid;
-    border-color: #ffffff;
-    border-width: 0px 1px 0px 0px;
-    width: 250.01px;
-    height: 84.38px;
-    position: absolute;
-    left: 0.99px;
-    top: 348.05px;
-  }
-  .map {
-    color: #000000;
-    text-align: center;
-    font-family: "Stylish-Regular", sans-serif;
-    font-size: 25px;
-    font-weight: 400;
-    position: absolute;
-    left: 0px;
-    top: 337.5px;
-    width: 248.38px;
-    height: 104.41px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    -webkit-text-stroke: 1px #ffffff;
-  }
-  .image-12 {
-    width: 37px;
-    height: 37px;
-    position: absolute;
-    left: 17px;
-    top: 351px;
-    object-fit: cover;
-    aspect-ratio: 1;
-  }
-  .image-13 {
-    width: 25px;
-    height: 25px;
-    position: absolute;
-    left: 23px;
-    top: 436px;
-    object-fit: cover;
-    aspect-ratio: 1;
-  }
-  .image-14 {
-    width: 26px;
-    height: 26px;
-    position: absolute;
-    left: 23px;
-    top: 507px;
-    object-fit: cover;
-    aspect-ratio: 1;
-  }
   .rectangle-12 {
     background: #d9d9d9;
     opacity: 0.8;
-    width: 1920px;
+    width: 1900px;
     height: 994px;
     position: absolute;
-    left: 0px;
     top: 86px;
   }
   .group-2 {
@@ -422,5 +183,5 @@
     transform-origin: 0 0;
     transform: rotate(89.819deg) scale(1, 1);
   }
-  </style>
+</style>
