+++ frontend/src/components_Back/NewPlaceComponent.vue
@@ -1,10 +1,21 @@
 <template>
-  <div class="full">
+  <div class="div">
+    <div class="rectangle-4"></div>
+    <!-- <img class="rectangle-9" src="/src/assets/images/public/rectangle-90.png" /> -->
+
+    <button class="btnLogout">logout</button>
     
+    <img @click="handleClick" class="image-15" src="/src/assets/images/public/image-150.png" />
+    <!-- <div class="rectangle-29"></div> -->
     <label><input type="text" class="rectangle-29"></label>
     <div class="place-name">Place Name</div>
+    <!-- <div class="rectangle-32"></div> -->
     <textarea class="rectangle-32"/>
     <div class="content">Content</div>
+
+    <!-- <div class="rectangle-23"></div>
+    <img class="polygon-5" src="/src/assets/images/newPlaceComponet/polygon-50.png" />
+    <img class="ellipse-2" src="/src/assets/images/public/ellipse-20.png" /> -->
 
     <!-- 색상 선택 드롭다운 -->
     <div class="dropdown-container" @click="toggleDropdown">
@@ -21,8 +32,14 @@
       </ul>
     </div>
 
+
+    <!-- <div class="rectangle-35"></div>
+    <div class="rectangle-36"></div> -->
+
     <button class="btnSave">save</button>
     <button class="btnCancel">cancel</button>
+
+
 
     <div class="rectangle-33"></div>
 
@@ -39,6 +56,25 @@
       <input id="photosInput" type="file" @change="addImages" multiple /> <br>
       <img v-for="(item, index) in src" :key="index" :src="item" />
     </label>
+
+    <div class="mappride">Mappride</div>
+    <!-- <div class="rectangle-6"></div> -->
+    <!-- <a href="#" class="my-categories">categories</a>
+    <a href="#" class="my-page">My page</a>
+    <a href="#" class="map">Map</a>
+    <img class="image-12" src="/src/assets/images/public/image-120.png" />
+    <img class="image-13" src="/src/assets/images/public/image-130.png" />
+    <img class="image-14" src="/src/assets/images/public/image-140.png" /> -->
+
+    <button class="map" @click="navigateTo('map')">
+          <img class="imgMap" src="/src/assets/images/public/image-120.png" />
+          Map</button>
+      <button class="my-page" @click="navigateTo('my-page')">
+          <img class="imgPage" src="/src/assets/images/public/image-130.png" />
+          My Page</button>
+      <button class="my-categories" @click="navigateTo('my-categories')">
+          <img class="imgCategory" src="/src/assets/images/public/image-140.png" />
+          My Categories</button>    
   </div>
 </template>
 <script>
@@ -101,17 +137,15 @@ export default {
 };
 </script>
 <style scoped>
-.full,
-.full * {
+.div,
+.div * {
   box-sizing: border-box;
 }
-.full {
-  margin-top: -950px;
-  margin-left: 300px;
+.div {
   background: #ffffff;
+  height: 1080px;
   position: relative;
   overflow: hidden;
-  height: 1080px;
 }
 .group-12 {
   width: 248.36px;
@@ -122,6 +156,17 @@ export default {
   width: 248.36px;
   height: 104.41px;
   position: static;
+}
+.rectangle-4 {
+  background: #ffffff;
+  border-style: solid;
+  border-color: #d3d3d3;
+  border-width: 0px 0px 1px 0px;
+  width: 1670px;
+  height: 90px;
+  position: absolute;
+  left: 250px;
+  top: 0px;
 }
 .rectangle-9 {
   width: 132px;
@@ -143,7 +188,24 @@ export default {
   width: 70px;
   height: 27px;
 } */
-
+.btnLogout {
+    color: #000000;
+    text-align: left;
+    font-family: "Stylish-Regular", sans-serif;
+    font-size: 20px;
+    font-weight: 400;
+    position: absolute;
+    left: 1804px;
+    top: 27px;
+    width: 120px;
+    height: 45px;
+    background-color: transparent; /* 배경을 투명하게 설정 */
+    border: 2px solid #d2d2d2; /* 테두리 추가 (두께 2px, 색상 #d2d2d2) */
+    cursor: pointer; /* 마우스를 올렸을 때 커서가 포인터로 바뀌게 설정 */
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
   .btnSave {
     color: #000000;
     text-align: left;
@@ -180,7 +242,15 @@ export default {
     align-items: center;
     justify-content: center;
   }
-
+.image-15 {
+  width: 46px;
+  height: 46px;
+  position: absolute;
+  left: 1624px;
+  top: 22px;
+  object-fit: cover;
+  aspect-ratio: 1;
+}
 .rectangle-29 {
   background: #f7f7f9;
   width: 244px;
@@ -429,6 +499,19 @@ export default {
   position: absolute;
   inset: 0;
 }
+.mappride {
+  color: #000000;
+  text-align: left;
+  font-family: "Stylish-Regular", sans-serif;
+  font-size: 30px;
+  font-weight: 400;
+  position: absolute;
+  left: 57.39px;
+  top: 31.64px;
+  width: 136.99px;
+  height: 40.08px;
+  -webkit-text-stroke: 1px #ffffff;
+}
 /* .rectangle-6 {
   background: #f7f7f7;
   border-style: solid;
@@ -488,6 +571,39 @@ export default {
   justify-content: center;
   -webkit-text-stroke: 1px #ffffff;
 } */
+.map, .my-page, .my-categories {
+    color: #000000;
+    text-align: center;
+    font-family: "Stylish-Regular", sans-serif;
+    font-size: 25px;
+    font-weight: 400;
+    position: absolute;
+    width: 248.36px;
+    height: 84.38px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    -webkit-text-stroke: 1px #ffffff;
+    background-color: transparent;
+    border: none; /* 버튼 테두리 */
+    cursor: pointer; /* 클릭 시 손 모양 커서 */
+  }
+  
+  .map {
+    top: 335.39px;
+    left: 20.14px;
+  }
+  
+  .my-page {
+    top: 439.80px;
+    left: 20.14px;
+  }
+  
+  .my-categories {
+    top:  544.21px;
+    left: 20.14px;
+  }
+  
   /* Hover 상태 */
   button:hover {
     background-color: #f0f0f0;
@@ -504,5 +620,32 @@ export default {
   position: absolute;
   left: 0px;
   top: 0px;
+}
+.image-12 {
+  width: 37.7px;
+  height: 37.7px;
+  position: absolute;
+  left: 16.09px;
+  top: 368.75px;
+  object-fit: cover;
+  aspect-ratio: 1;
+}
+.image-13 {
+  width: 25.47px;
+  height: 25.47px;
+  position: absolute;
+  left: 22.21px;
+  top: 457.13px;
+  object-fit: cover;
+  aspect-ratio: 1;
+}
+.image-14 {
+  width: 26.49px;
+  height: 26.49px;
+  position: absolute;
+  left: 22.19px;
+  top: 534.14px;
+  object-fit: cover;
+  aspect-ratio: 1;
 }
 </style>
